// client/da/uml-server/src/language-server/generated/ast.ts
export type Entity = Class | DataType | Enumeration | InstanceSpecification | Interface | Package | PrimitiveType;

export const Entity = 'Entity';
// ...
export type SlotDefiningFeature = Class | Interface | Property;

export const SlotDefiningFeature = 'SlotDefiningFeature';

export function isSlotDefiningFeature(item: unknown): item is SlotDefiningFeature {
    return reflection.isInstance(item, SlotDefiningFeature);
}

export type UnionType_0 = ClassDiagram | PackageDiagram | StateMachineDiagram;

export const UnionType_0 = 'UnionType_0';

export function isUnionType_0(item: unknown): item is UnionType_0 {
    return reflection.isInstance(item, UnionType_0);
}

export type Visibility = 'PACKAGE' | 'PRIVATE' | 'PROTECTED' | 'PUBLIC';
// ...
export interface ClassDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'ClassDiagram';
    __id: string
    diagramType: 'CLASS'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const ClassDiagram = 'ClassDiagram';
// ...
export interface Diagram extends AstNode {
    readonly $type: 'Diagram';
    diagram: UnionType_0
    metaInfos: Array<MetaInfo>
}

export const Diagram = 'Diagram';
// ...
export interface InstanceSpecification extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'InstanceSpecification';
    __id: string
    name: string
    slots: Array<Slot>
    visibility?: Visibility
}

export const InstanceSpecification = 'InstanceSpecification';
// ...
export interface LiteralSpecification extends AstNode {
    readonly $container: Slot;
    readonly $type: 'LiteralSpecification';
    __id: string
    name: string
    value: string
}

export const LiteralSpecification = 'LiteralSpecification';
// ...
export interface Package extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Package';
    __id: string
    entities: Array<Entity>
    name: string
    uri?: string
    visibility?: Visibility
}

export const Package = 'Package';

export function isPackage(item: unknown): item is Package {
    return reflection.isInstance(item, Package);
}

export interface PackageDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'PackageDiagram';
    __id: string
    diagramType: 'PACKAGE'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const PackageDiagram = 'PackageDiagram';
// ...
export interface Relation extends AstNode {
    readonly $type: 'Abstraction' | 'Aggregation' | 'Association' | 'Composition' | 'Dependency' | 'Generalization' | 'InterfaceRealization' | 'PackageImport' | 'PackageMerge' | 'Realization' | 'Relation' | 'Substitution' | 'Usage';
    __id: string
    relationType: RelationType
    source: Reference<Entity>
    target: Reference<Entity>
}

export const Relation = 'Relation';
// ...
export interface Slot extends AstNode {
    readonly $container: InstanceSpecification;
    readonly $type: 'Slot';
    __id: string
    definingFeature?: Reference<SlotDefiningFeature>
    name: string
    values: Array<LiteralSpecification>
}

export const Slot = 'Slot';